requirements.txt
numpy>=2.1.3
soundfile>=0.13.1
mido>=1.2.10
Flask>=2.3.3
pyrubberband>=0.3.0
audiotsm>=0.1.2
librosa>=0.10.2.post1
requests>=2.31.0
pyserial>=3.5
flask-socketio>=5.3.0
python-socketio>=5.9.0
eventlet>=0.33.0
